mercury_demo/mercury_demo_requirements.txt
transformers
torch
mljar-mercury
pip
